Conflicts:
	Projet_ABD3/src/Connexionsgbd.java
	Projet_ABD3/src/InterfaceClient.java

+++ Projet_ABD3/src/Interactions/InterfaceGestion.java
@@ -74,7 +74,7 @@ public class InterfaceGestion {
 		
 		st.executeQuery();
 		conn.commit();
-		System.out.println("Dispositif" + idD + "  ajouté pour la société "+societe);
+		System.out.println("Dispositif " + idD + "  ajouté pour la société "+societe);
 		
 		return false;
 	}
@@ -96,14 +96,17 @@ public class InterfaceGestion {
 		conn.commit();
 		
 		PreparedStatement st = conn.prepareStatement("insert into FormatSociete values (?,?,?,?,?)");
-		st.setInt(1, idS);
-		st.setInt(2, idF);
+		st.setInt(1, idF);
+		st.setInt(2, idS);
 		st.setInt(3, stock);
 		st.setFloat(4, prixU);
 		st.setInt(5, tirageJour);
 		st.executeQuery();
 		
 		conn.commit();
+		
+		System.out.println("Format " + idF + "  ajouté pour la societé : " + societe);
+
 		
 		return false;
 	}
@@ -129,30 +132,39 @@ public class InterfaceGestion {
 
 	 public void supprimerDispositif(int idD) throws SQLException {
 		 
-		 
-		 PreparedStatement req = conn.prepareStatement("select idF from FormatSociete natural join DispositifFormat where idD=?");
-			req.setInt(1,idD);
-			ResultSet res = req.executeQuery();
+		 PreparedStatement req0 = conn.prepareStatement("select count(idAlbum) from commande natural join dispositif natural join lot where idD=? and (statutCommande='en cours' or statutCommande='envoi partiel' )");
+			req0.setInt(1,idD);
+			ResultSet res0 = req0.executeQuery();
+			res0.next();
+			int nbCom = res0.getInt(1);
 			
+			if (nbCom==0) {
 			
-			while(res.next()) {
-				int idF = res.getInt(1);
-				PreparedStatement req2 = conn.prepareStatement("delete from FormatSociete where idF=?");
-				req2.setInt(1, idF);
-				req2.executeQuery();			
-			}
-		 
-			PreparedStatement req3 = conn.prepareStatement("delete from DispositifFormat where idD=?");
-			req3.setInt(1, idD);
-			req3.executeQuery();
-			
-			PreparedStatement req4 = conn.prepareStatement("delete from dispositif where idD=?");
-			req4.setInt(1, idD);
-			req4.executeQuery();
-			conn.commit();
-			
-			System.out.println("Dispositif" + idD + "  supprimé.");
+					
+				 PreparedStatement req = conn.prepareStatement("select idF from FormatSociete natural join DispositifFormat where idD=?");
+					req.setInt(1,idD);
+					ResultSet res = req.executeQuery();
+					
+					
+					while(res.next()) {
+						int idF = res.getInt(1);
+						PreparedStatement req2 = conn.prepareStatement("delete from FormatSociete where idF=?");
+						req2.setInt(1, idF);
+						req2.executeQuery();			
+					}
+				 
+					PreparedStatement req3 = conn.prepareStatement("delete from DispositifFormat where idD=?");
+					req3.setInt(1, idD);
+					req3.executeQuery();
+					
+					PreparedStatement req4 = conn.prepareStatement("delete from dispositif where idD=?");
+					req4.setInt(1, idD);
+					req4.executeQuery();
+					conn.commit();
+					
+					System.out.println("Dispositif" + idD + "  supprimé.");
 
+			}
 			/*
 			 * Trigger : verifier si dernier dispositif a faire un format pour une societe
 			 */
@@ -172,7 +184,7 @@ public class InterfaceGestion {
 			int idS = res.getInt(1);
 			req.close();
 			
-			PreparedStatement req2 = conn.prepareStatement("select count(idLot) from Lot where idS=?");
+			PreparedStatement req2 = conn.prepareStatement("select count(idLot) from Lot natural join commande where idS=? and(statutcommande='en cours' or statutcommande='envoi partiel')");
 			req2.setInt(1,idS);
 			ResultSet res2 = req2.executeQuery();
 			res2.next();
@@ -180,10 +192,22 @@ public class InterfaceGestion {
 			req2.close();
 			
 			if (nbLot == 0) {
+				
+				 PreparedStatement req4 = conn.prepareStatement("select idD from dispositif where idS=?");
+					req4.setInt(1,idS);
+					ResultSet res4 = req4.executeQuery();		
+					
+					while(res4.next()) {
+						int idD = res4.getInt(1);
+						supprimerDispositif(idD);				
+					}
+					req4.close();
+							
 				PreparedStatement req3 = conn.prepareStatement("delete from societe where idS=?");
 				req3.setInt(1, idS);
 				req3.executeQuery();
 				conn.commit();
+				
 				
 				System.out.println("Prestataire " + societe + "  supprimé. ");
 
@@ -210,7 +234,7 @@ public class InterfaceGestion {
 		  * A COMPLETER !
 		  */	 
 		 
-			PreparedStatement req2 = conn.prepareStatement("select count(idCom) from commande where mailClient=? and (statutcommande='envoi partiel' or statutcommande='en cours' )");
+			PreparedStatement req2 = conn.prepareStatement("select count(idCom) from commande where mailClient=? and (statutcommande='envoie partiel' or statutcommande='en cours' )");
 			req2.setString(1,mail);
 			ResultSet res2 = req2.executeQuery();
 			res2.next();
@@ -232,18 +256,24 @@ public class InterfaceGestion {
 						req1.setInt(1,idCom);
 						ResultSet res3 = req5.executeQuery();
 								while(res3.next()) {
+										int idLot =res3.getInt(1);
 										PreparedStatement req8 = conn.prepareStatement("delete from livraison where idLot=?");
-										req8.setInt(1, idCom);
+										req8.setInt(1, idLot);
 										req8.executeQuery();	
 								}
+							System.out.println("loool");
 							
-						PreparedStatement req4 = conn.prepareStatement("delete from Lot where idCom=?");
-						req4.setInt(1, idCom);
-						req4.executeQuery();	
-
-						PreparedStatement req6 = conn.prepareStatement("delete from commande where idCom=?");
-						req6.setInt(1, idCom);
-						req6.executeQuery();						
+							
+										PreparedStatement req9 = conn.prepareStatement("delete from Lot where idCom=?");
+										req9.setInt(1, idCom);
+										req9.executeQuery();	
+								
+	
+										PreparedStatement req6 = conn.prepareStatement("delete from commande where idCom=?");
+										req6.setInt(1, idCom);
+										req6.executeQuery();	
+									
+					
 					}
 				/*
 				PreparedStatement req3 = conn.prepareStatement("delete from Client where mailClient=?");
@@ -265,6 +295,8 @@ public class InterfaceGestion {
 				*/
 			}
 			else {
+				System.out.println("laaaal");
+
 				PreparedStatement st2 = conn.prepareStatement("insert into ListeSuppClient values (?,?,?)");
 				st2.setString(1, mail);
 				st2.setInt(2, nbCom);
@@ -279,7 +311,7 @@ public class InterfaceGestion {
 				System.out.println("Client: " + mail + "  ajouté a la file d'attente de suppression. ");
 				
 				
-				 PreparedStatement req5 = conn.prepareStatement("select idI from Images where mailclient=? and ");
+				 PreparedStatement req5 = conn.prepareStatement("select idI from Image where mailclient=? ");
 					req5.setString(1,mail);
 					ResultSet res = req5.executeQuery();
 					
@@ -290,7 +322,7 @@ public class InterfaceGestion {
 						
 						PreparedStatement req1 = conn.prepareStatement("update image set partage=0 where idI=?");
 						req1.setInt(1,idI);
-						ResultSet res3 = req1.executeQuery();		
+						req1.executeQuery();		
 						supprimmerPhotosSansCommande(idI);
 						//PreparedStatement req8 = conn.prepareStatement("delete from photo where idI=?");
 						//req8.setInt(1,idI);
@@ -301,19 +333,28 @@ public class InterfaceGestion {
 		 
 	 }
 	 
-	 private void supprimmerPhotosSansCommande(int idI) throws SQLException {
+	 private void supprimmerPhotosSansCommande(int idI) throws SQLException  {
 		
 		 PreparedStatement req1 = conn.prepareStatement("select idPhoto from"
-		 		+ " photo join Album using(idPhoto) join lot using(idAlbum) join commande using(idCom)"
+		 		+ " photo join Album using(idAlbum) join lot using(idAlbum) join commande using(idCom)"
 		 		+ " where idI=? and not(statutCommande='en cours' or statutCommande='envoi partiel' )");
 			req1.setInt(1,idI);
 			ResultSet res = req1.executeQuery();
 			while(res.next()) {
 				int idPhoto = res.getInt(1);
-
-				PreparedStatement req6 = conn.prepareStatement("delete from photo where idPhoto=?");
-				req6.setInt(1, idPhoto);
-				req6.executeQuery();						
+				
+				try {
+					PreparedStatement req6;
+					req6 = conn.prepareStatement("delete from photo where idPhoto=?");
+					req6.setInt(1, idPhoto);
+					req6.executeQuery();		
+				} catch (SQLException e) {
+					PreparedStatement req7;
+					req7 = conn.prepareStatement("update photo set idI=0 where idPhoto=?");
+					req7.setInt(1, idPhoto);
+					req7.executeQuery();	
+				}
+				
 			}
 			insererImageListeSuppp(idI);
